moved imports around, replaced bcrypt with js-only version for buildability
