+++ api/babel.config.json
@@ -1,0 +1,11 @@
+{
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "useBuiltIns": "usage",
+        "corejs": "2.6.11" // match this to package.json version
+      }
+    ]
+  ]
+}
