+++ api/package.json
@@ -2,10 +2,12 @@
   "name": "api",
   "version": "1.0.0",
   "description": "REST API for my personal website",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index.js"
+    "start": "node dist/index.js",
+    "pre:start": "npm run build",
+    "build": "./node_modules/.bin/babel src --out-dir dist"
   },
   "repository": {
     "type": "git",
@@ -18,22 +20,26 @@
   },
   "homepage": "https://github.com/PeterSherk/vue-website#readme",
   "dependencies": {
-    "@babel/polyfill": "^7.12.1",
-    "bcrypt": "^5.0.1",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "helmet": "^3.23.3",
-    "jsonwebtoken": "^8.5.1",
-    "openapi-types": "^7.2.3",
-    "pg": "^7.18.2",
-    "swagger-jsdoc": "^4.3.2",
-    "swagger-ui-express": "^4.1.6",
-    "winston": "^3.3.3"
+    "bcryptjs": "2.4.3",
+    "core-js": "2.6.11",
+    "cors": "2.8.5",
+    "dotenv": "10.0.0",
+    "express": "4.17.1",
+    "helmet": "4.6.0",
+    "install": "^0.13.0",
+    "jsonwebtoken": "8.5.1",
+    "npm": "^8.1.3",
+    "openapi-types": "9.3.1",
+    "pg": "8.7.1",
+    "regenerator-runtime": "0.13.9",
+    "swagger-jsdoc": "6.1.0",
+    "swagger-ui-express": "4.1.6",
+    "winston": "3.3.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "nodemon": "^2.0.7"
+    "@babel/cli": "7.16.0",
+    "@babel/core": "7.16.0",
+    "@babel/preset-env": "7.16.0",
+    "nodemon": "2.0.15"
   }
 }
