+++ api/src/index.js
@@ -1,0 +1,3 @@
+import 'core-js/es6';
+import 'regenerator-runtime/runtime';
+import "./app/router"
